+++ app/google-services.json
@@ -1,0 +1,42 @@
+{
+  "project_info": {
+    "project_number": "913926382626",
+    "firebase_url": "https://android-nanodegree-sunshine.firebaseio.com",
+    "project_id": "android-nanodegree-sunshine",
+    "storage_bucket": "android-nanodegree-sunshine.appspot.com"
+  },
+  "client": [
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:913926382626:android:83b792e26f07bf18",
+        "android_client_info": {
+          "package_name": "com.example.android.sunshine.app"
+        }
+      },
+      "oauth_client": [
+        {
+          "client_id": "913926382626-cfet7ve9s1r5d4223i3k0qect8bbce82.apps.googleusercontent.com",
+          "client_type": 3
+        }
+      ],
+      "api_key": [
+        {
+          "current_key": "AIzaSyDlhAhRNsL9LImd_jW689akFpstNLfKzHg"
+        }
+      ],
+      "services": {
+        "analytics_service": {
+          "status": 1
+        },
+        "appinvite_service": {
+          "status": 1,
+          "other_platform_oauth_client": []
+        },
+        "ads_service": {
+          "status": 2
+        }
+      }
+    }
+  ],
+  "configuration_version": "1"
+}
