--- app/src/main/java/com/example/android/sunshine/app/gcm/MyInstanceIDListenerService.java
@@ -1,36 +1,0 @@
-/*
- * Copyright (C) 2015 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.example.android.sunshine.app.gcm;
-
-import android.content.Intent;
-import com.google.android.gms.iid.InstanceIDListenerService;
-
-public class MyInstanceIDListenerService extends InstanceIDListenerService {
-    private static final String TAG = "MyInstanceIDLS";
-
-    /**
-     * Called if InstanceID token is updated. This may occur if the security of
-     * the previous token had been compromised. This call is initiated by the
-     * InstanceID provider.
-     */
-    @Override
-    public void onTokenRefresh() {
-        // Fetch updated Instance ID token.
-        Intent intent = new Intent(this, RegistrationIntentService.class);
-        startService(intent);
-    }
-}
