--- app/src/main/java/com/example/android/sunshine/app/gcm/RegistrationIntentService.java
@@ -1,82 +1,0 @@
-/*
- * Copyright (C) 2015 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.example.android.sunshine.app.gcm;
-
-import android.app.IntentService;
-import android.content.Intent;
-import android.content.SharedPreferences;
-import android.preference.PreferenceManager;
-import android.util.Log;
-import android.widget.Toast;
-
-import com.example.android.sunshine.app.MainActivity;
-import com.example.android.sunshine.app.R;
-import com.google.android.gms.gcm.GoogleCloudMessaging;
-import com.google.android.gms.iid.InstanceID;
-
-
-public class RegistrationIntentService extends IntentService {
-    private static final String TAG = "RegIntentService";
-
-    public RegistrationIntentService() {
-        super(TAG);
-    }
-
-    @Override
-    protected void onHandleIntent(Intent intent) {
-        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
-
-        try {
-            // In the (unlikely) event that multiple refresh operations occur simultaneously,
-            // ensure that they are processed sequentially.
-            synchronized (TAG) {
-                // Initially this call goes out to the network to retrieve the token, subsequent calls
-                // are local.
-                InstanceID instanceID = InstanceID.getInstance(this);
-
-                // TODO: gcm_default sender ID comes from the API console
-                String senderId = getString(R.string.gcm_defaultSenderId);
-                if ( senderId.length() != 0 ) {
-                    String token = instanceID.getToken(senderId,
-                            GoogleCloudMessaging.INSTANCE_ID_SCOPE, null);
-                    sendRegistrationToServer(token);
-                }
-
-                // You should store a boolean that indicates whether the generated token has been
-                // sent to your server. If the boolean is false, send the token to your server,
-                // otherwise your server should have already received the token.
-                sharedPreferences.edit().putBoolean(MainActivity.SENT_TOKEN_TO_SERVER, true).apply();
-            }
-        } catch (Exception e) {
-            Log.d(TAG, "Failed to complete token refresh", e);
-
-            // If an exception happens while fetching the new token or updating our registration data
-            // on a third-party server, this ensures that we'll attempt the update at a later time.
-            sharedPreferences.edit().putBoolean(MainActivity.SENT_TOKEN_TO_SERVER, false).apply();
-        }
-    }
-
-    /**
-     * Normally, you would want to persist the registration to third-party servers. Because we do
-     * not have a server, and are faking it with a website, you'll want to log the token instead.
-     * That way you can see the value in logcat, and note it for future use in the website.
-     *
-     * @param token The new token.
-     */
-    private void sendRegistrationToServer(String token) {
-        Log.i(TAG, "GCM Registration Token: " + token);
-    }
-}
